Move fps view controller
